requirements/base.txt
Django==3.2.12
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
psycopg2==2.9.3
requests==2.27.1
